### 2022-04-20 #4332 @arfon
:wave: @yuliadm  – thanks for your submission to JOSS. While I look for an editor, please work to redraft and reduce the size of your paper. 1000 words us our soft limit and there appears to be a lot written here. We have detailed guidance here: https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain
