+++ paper.md
@@ -27,7 +27,7 @@ bibliography: refs.bib
 
 # Summary
 
-The **mixComp** package provides a number of methods for obtaining a consistent estimate of the complexity of a finite mixture (the focus is made on the univariate case). The considered approaches can be loosely grouped into three categories:
+The **mixComp** package provides several methods for obtaining a consistent estimate of the complexity of a (univariate) finite mixture. The considered approaches can be loosely grouped into three categories:
 
   - methods built upon the determinants of the Hankel matrix of moments of the mixing distribution; 
   
@@ -35,31 +35,13 @@ The **mixComp** package provides a number of methods for obtaining a consistent 
    
   - likelihood ratio test (LRT) - based techniques. 
 
-While not the primary goal, most methods simultaneously estimate the component weights and parameters. In this document, we give a brief overview of the methodology, and demonstrate the package's functionality in both real world examples and synthetically generated data. Moreover, we show how the package can be used on virtually any parametric mixture as long as functions generating random variates and evaluating the density are provided for the component distribution.
+While not the primary goal, most methods simultaneously estimate the component weights and parameters. 
 
 # Statement of need
 
-Mixture models occur in numerous settings including random
-and fixed effects models, clustering, deconvolution, empirical Bayes problems 
-and many others. In particular, they are often used to model the data
-which are believed to come from a heterogeneous population, consisting of
-several homogeneous subpopulations. In such cases the problem of finding
-a good estimator for the number of components in the mixture arises naturally, 
-whether of primary interest in itself or as a first step in estimating the
-population distribution. Estimation of the order of a finite mixture model
-is known to be a hard statistical task, and multiple techniques have been
-suggested for solving it.
+**mixComp** can be used on virtually any parametric mixture as long as functions generating random variates and evaluating the density are provided for the component distribution. IT is applicable to parametric mixtures well beyond those whose component distributions are included in the **R** package **stats**, making it more customizable than most packages for model-based clustering. The estimation results can be printed out and plotted for further analysis and study.
 
-The **mixComp** package provides a variety of approaches for estimation of the complexity of
-a finite mixture distribution, simultaneously estimating the component weights and parameters. 
-It is applicable to parametric mixtures well beyond those whose component distributions are included in the **R** package **stats**, making it more customizable than most packages for model-based clustering. 
-The estimation results can be printed out and plotted for further analysis and study.
-The documentation contains multiple examples based on simulated as well as real data, several real-world 
-datasets have been built-in to the package for convenience.
-
-The use of the **mixComp** package might be of interest to researchers and practitioners who 
-are studying phenomena that can be effectively modelled using mixture distributions. 
-Among other things it can be used to identify settings and conditions, under which a certain 
+The use of the **mixComp** package might be of interest to researchers and practitioners who are studying phenomena that can be effectively modelled using mixture distributions. In particular, it can be used to identify settings and conditions, under which a certain 
 method provides more accurate estimates than the others.
 
 # Installation
@@ -70,11 +52,11 @@ install.packages("mixComp")
 ```
 # Section 1. Introduction to finite mixture models and mixComp
 
-Mixture models have been used extensively in statistical applications and therefore have attracted a lot of attention from both theoretical and computational perspectives. Although the list of works on mixture models is too long to make an exhaustive inventory, we can refer to the following important papers and books: [@Teicher63], [@LindsayI], [@LindsayII], [@Titterington] and [@McLachlan].
[…]
