+++ paper.md
@@ -136,7 +136,7 @@ plot(poisRMix, main = "3-component poisson mixture", cex.main = 0.9)
 
 ![3-component poisson mixture](figures/poisRMix.png) 
   
-The third object class shown in Table 1, called `datMix`, represents the data vector $\mathbf{X}$ based on which the mixture complexity is supposed to be estimated. These objects are most central to the package, as every procedure estimating the order of a mixture takes a `datMix` object as input. Apart from $\mathbf{X}$, it contains other 'static' information needed for the estimation procedure in contrast to 'tuning parameters' which can be changed with every function call (e.g. the number of bootstrap replicates for a function employing a bootstrap procedure). An overview of which "static" attributes need to be supplied for each complexity estimation routine is given in Table 2. 
+The third object class shown in Table 1, called `datMix`, represents the data vector $\mathbf{X}$ based on which the mixture complexity should be estimated. These objects are key, as every procedure estimating the order of a mixture takes a `datMix` object as input. Apart from $\mathbf{X}$, it contains other 'static' information needed for the estimation procedure in contrast to 'tuning parameters' which can be changed with every function call. An overview of which 'static' attributes need to be supplied for each complexity estimation routine is given in Table 2. 
 
 #### Table 2: Inputs needed for different functions contained in mixComp
 | `R function`            |`dist`|`discrete`|`theta.bound.list`|  `MLE.function` | `Hankel.method` | `Hankel.function` |
@@ -150,22 +150,22 @@ The third object class shown in Table 1, called `datMix`, represents the data ve
 
 In the table, "o" and "i" stand for input used during optimization and initialization.
 
-As an example consider the `waiting` from the Old Faithful dataset [@R; @faithful1; @faithful2]. To estimate the number of components of the mixture distribution that provides a suitable approximation to the `waiting` data via **mixComp**, the raw data vector of observations has to be converted to a `datMix` object first. For the sake of exposition we specify all arguments of  the `datMix` function, starting with the vector of observations $\mathbf{X}$ and the string `dist`, specifying $\{g(x;\theta):\theta \in \Theta \}$ and the boolean `discrete`. As has often been done in the relevant literature, we assume that the data comes from a normal mixture.
+As an example consider the `waiting` (often assumed to be generated by a normal mixture) from the Old Faithful dataset [@R; @faithful1; @faithful2]. To estimate the number of components of the mixture distribution that provides a suitable approximation to the `waiting` via **mixComp**, the raw data vector of observations $\mathbf{X}$ should be converted to a `datMix` object. Then arguments of the `datMix` function are entered, starting with $\mathbf{X}$ and the string `dist`, specifying $\{g(x;\theta):\theta \in \Theta \}$ and the boolean `discrete`.
 
 ``` r
 faithful.obs <- faithful$waiting
 norm.dist <- "norm"
 norm.discrete <- FALSE
 ```
-Second, a named list of length $d$ containing the bounds of $\theta \in \Theta \subseteq \mathbf{R}^d$ has to be created. In this example, $\theta = \{\mu, \sigma\} \in \Theta = \mathbf{R} \times (0, \infty) \subseteq \mathbf{R}^2$.
+Now a named list of length $d$ containing the bounds of $\theta \in \Theta \subseteq \mathbf{R}^d$ has to be created. In this example, $\theta = \{\mu, \sigma\} \in \Theta = \mathbf{R} \times (0, \infty) \subseteq \mathbf{R}^2$.
 
 ``` r
 # define the range for parameter values:
 norm.bound.list <- list("mean" = c(-Inf, Inf), "sd" = c(0, Inf))
[…]
