+++ paper.md
@@ -25,7 +25,7 @@ bibliography: refs.bib
 
 # Summary
 
-The **mixComp** package provides several methods for obtaining a consistent estimate of the complexity of a (univariate) finite mixture. The considered approaches can be loosely grouped into three categories:
+The **mixComp** package provides several methods for obtaining a consistent estimate of the complexity of a (univariate) finite mixture. The considered approaches can be loosely grouped into 3 categories:
 
   - methods built upon the determinants of the Hankel matrix of moments of the mixing distribution; 
   
@@ -37,10 +37,9 @@ While not the primary goal, most methods simultaneously estimate the component w
 
 # Statement of need
 
-**mixComp** can be used on virtually any parametric mixture as long as functions generating random variates and evaluating the density are provided for the component distribution. IT is applicable to parametric mixtures well beyond those whose component distributions are included in the **R** package **stats**, making it more customizable than most packages for model-based clustering. The estimation results can be printed out and plotted for further analysis and study.
+**mixComp** can be used on virtually any parametric mixture as long as functions generating random variates and evaluating the density are provided for the component distribution. It is applicable to parametric mixtures well beyond those whose component distributions are included in the **R** package **stats**, making it more customizable than most packages for model-based clustering. The estimation results can be printed out and plotted for further analysis.
 
-The use of the **mixComp** package might be of interest to researchers and practitioners who are studying phenomena that can be effectively modelled using mixture distributions. In particular, it can be used to identify settings and conditions, under which a certain 
-method provides more accurate estimates than the others.
+The package is aimed at practitioners studying phenomena that can be effectively modelled using mixture distributions. In particular, it can be used to identify settings and conditions, under which a certain method provides more accurate estimates than the others.
 
 # Installation
 
@@ -48,11 +47,12 @@ To install from `CRAN`, use:
 ``` r
 install.packages("mixComp")
 ```
+
 # Section 1. Introduction to finite mixture models and mixComp
 
 Mixture models have been used extensively in statistical applications and have attracted much attention from both theoretical and computational perspectives. The list of works on mixture models is extensive; we can refer to the following papers and books: [@Teicher63], [@LindsayI], [@LindsayII], [@Titterington] and [@McLachlan].
 
[…]
