+++ paper.md
@@ -52,7 +52,7 @@ install.packages("mixComp")
 
 Mixture models have been used extensively in statistical applications and have attracted much attention from both theoretical and computational perspectives. The list of works on mixture models is extensive; we can refer to the following papers and books: [@Teicher63], [@LindsayI], [@LindsayII], [@Titterington] and [@McLachlan].
 
-The popularity of such models stems from the fact that they allow for modeling heterogeneous data whose distribution cannot be captured by a single parametric distribution. To account for such heterogeneity, the (unknown) distribution is assumed to result from mixing over some latent parameter in the following sense: the latent parameter is viewed itself as a random variable drawn from some unknown mixing distribution. When this mixing distribution is only assumed to belong to the ensemble of all possible distribution functions, the mixture model is called *nonparametric* and estimation of the mixing distribution requires using some nonparametric estimation method. This includes the well-known nonparametric maximum likelihood estimator (NPMLE) whose fundamental properties were well studied in the seminal work of [@LindsayI], [@LindsayII]. One remarkable property of the NPMLE of the mixing distribution is that it is, under some simple conditions, a discrete distribution function with at most $k$ number of jumps, where $k$ is the number of distinct observations in the random sample drawn from the mixture. This interesting feature is one reason for considering the smaller class of finite mixture models (characterized by a discrete mixing distribution with a finite number of jumps). The model has the following simple interpretation: the population under study is assumed to consist of several homogeneous subpopulations. These subpopulations, typically referred to as the mixture's components, usually have a specific meaning depending on the problem at hand. In some very simple situations, the number of components could be known in advance, in which case the model is *fully parametric* and convergence of classical estimators such as the parametric maximum likelihood estimator (MLE) is known to occur at the fast rate $\sqrt{n}$ (under some conditions). Also, the well-known expectation-maximization (EM) algorithm [@Dempster] can be used to find the MLE of all the unknown parameters. However, in many applications such knowledge is rarely available and the number of components has to be estimated from the data. Although the mixture is still finite and the distribution of each component is assumed to belong to some parametric family, the estimation framework in this case is much harder. The goal of **mixComp** is to estimate the unknown complexity using several methods known from the statistical literature. These methods, discussed below, all come with theoretical guarantees for consistency as the sample size gets larger. The performance of the methods varies according to the underlying mixture distribution and the sample size. This will be illustrated below through several synthetic and real datasets.
[…]
