+++ paper.md
@@ -29,7 +29,7 @@ The **mixComp** package provides several methods for obtaining a consistent esti
 
   - methods built upon the determinants of the Hankel matrix of moments of the mixing distribution; 
   
-  - methods based on penalized minimum distance between the unknown probability density and a consistent estimator thereof (the distances considered in this survey are the Hellinger and the $L_2$-distances);
+  - methods based on penalized minimum distance between the unknown probability density and a consistent estimator thereof;
    
   - likelihood ratio test (LRT) - based techniques. 
 
@@ -37,7 +37,7 @@ While not the primary goal, most methods simultaneously estimate the component w
 
 # Statement of need
 
-**mixComp** can be used on virtually any parametric mixture as long as functions generating random variates and evaluating the density are provided for the component distribution. It is applicable to parametric mixtures well beyond those whose component distributions are included in the **R** package **stats**, making it more customizable than most packages for model-based clustering. The estimation results can be printed out and plotted for further analysis. The package is aimed at practitioners studying phenomena that can be effectively modelled using mixture distributions. In particular, it can be used to identify settings and conditions, under which a certain method provides more accurate estimates than the others.
+**mixComp** can be used on virtually any parametric mixture as long as functions generating random variates and evaluating the density are provided for the component distribution. It is applicable to parametric mixtures well beyond those whose component distributions are included in the **R** package **stats**, making it more customizable than most packages for model-based clustering. The estimation results can be printed and plotted for further analysis. The package is aimed at practitioners studying phenomena that can be effectively modelled using mixture distributions. In particular, it can be used to identify settings and conditions, under which a certain method provides more accurate estimates than the others.
 
 # Installation
 
@@ -48,9 +48,9 @@ install.packages("mixComp")
 
 # Section 1. Introduction to finite mixture models and mixComp
 
-Mixture models have been used extensively in statistical applications and have attracted much attention from both theoretical and computational perspectives. The list of works on mixture models is extensive; we can refer to the following papers and books: [@Teicher63], [@LindsayI], [@LindsayII], [@Titterington] and [@McLachlan].
+Mixture models have been used extensively in statistical applications and have attracted much attention from theoretical and computational perspectives. The list of works on mixture models is extensive: [@Teicher63], [@LindsayI], [@LindsayII], [@Titterington], [@McLachlan], to name a few. 
 
[…]
