+++ README.md
@@ -1,0 +1,30 @@
+---
+title: 'mixComp: An R package for estimating complexity of a mixture'
+tags:
+  - R
+  - mixture distribution
+  - complexity
+authors:
+  - name: Anja Weigel^[Co-first author] # note this makes a footnote saying 'Co-first author'
+    affiliation: 2 # (Multiple affiliations must be quoted)
+  - name: Fadoua Balabdaoui^[Co-first author] 
+    affiliation: 1
+  - name: Yulia Kulagina^[Corresponding author] 
+    affiliation: 1
+  - name: Lilian Mueller^[Contributor]
+    affiliation: 2
+affiliations:
+ - name: ETH Zurich, Seminar for Statistics, Switzerland
+   index: 1
+ - name: ETH Zurich, Switzerland
+   index: 2
+date: 19 April 2022
+bibliography: refs.bib
+
+---
+
+# Installation
+
+To install from `CRAN`, use:
+```
+install.packages("mixComp")
