+++ README.md
@@ -20,7 +20,6 @@ affiliations:
    index: 2
 date: 19 April 2022
 bibliography: refs.bib
-
 ---
 
 # Installation
