+++ README.md
@@ -152,7 +152,7 @@ normLoc.dM <- RtoDat(normLocRMix, theta.bound.list = norm.bound.list,
 ```
 
 In the case of the scaled version of the method, the penalty should be scaled accordingly as well as mentioned earlier. 
-```{r nonph}
+``` r
 # define the penalty function:
 pen <- function(j, n){
   j * log(n)
@@ -168,7 +168,7 @@ normdets_sca_pen <- nonparamHankel(normLoc.dM, j.max = 5, scaled = TRUE,
 
 We can print and plot the results as suggested below.
 
-```{r plotnonph, figures-side, fig.show="hold", out.width="50%"}
+``` r
 # print the results (for the Poisson mixture)
 print(poisdets_sca_pen)
 # plot results for both mixtures:
@@ -188,20 +188,20 @@ Having created the data ourselves, we know that it comes from a 3-component Pois
 
 As a simple example of a given dataset to which mixture models have been applied extensively, take the Old Faithful dataset [[29]](#29), [[1]](#1), [[18]](#18). In the context of mixture model estimation, the variable `waiting`, which gives the time in minutes between eruptions of the Old Faithful geyser in the Yellowstone National Park, is often considered to be the variable of interest. To estimate the number of components of the mixture distribution that provides a suitable approximation to the `waiting` data via **mixComp**, the raw data vector of observations has to be converted to a `datMix` object first. For the sake of exposition we specify all arguments of  the `datMix` function. As has often been done in the relevant literature, we assume that the data comes from a normal mixture.
 
-```{r faithopts}
+``` r
 faithful.obs <- faithful$waiting
 norm.dist <- "norm"
 norm.discrete <- FALSE
 ```
 Now a named list containing the bounds for the component parameters (the mean and standard deviation) has to be created:
 
-```{r normlist}
+``` r
 # define the range for parameter values:
 norm.bound.list <- list("mean" = c(-Inf, Inf), "sd" = c(0, Inf))
 ```
 Next, the argument `MLE.function` for the mean and for the standard deviation have to be defined:
 
-```{r normfun}
+``` r
 # define the MLE functions for the mean and sd: 
 MLE.norm.mean <- function(dat) mean(dat)
 MLE.norm.sd <- function(dat){
@@ -211,7 +211,7 @@ MLE.norm.list <- list("MLE.norm.mean" = MLE.norm.mean, "MLE.norm.sd" = MLE.norm.
 ```
 The last two arguments, `Hankel.method` and `Hankel.function`, need to be supplied if the mixture complexity is to be estimated based on the Hankel matrix of the moments of the mixing distribution. The reader is referred to the Section 3 for further information on how these arguments are to be specified (in this case, the simplifying assumption of unit variance is made. This would be a poor choice for the `waiting` data, so number of mixture components should not be estimated with one of the methods using these arguments, namely `nonparamHankel`, `paramHankel` and `paramHankel.scaled`, see Table 2). 
 
-```{r normmom}
+``` r
 method <- "translation"
 # define the function for computing the moments:
 mom.std.norm <- function(j){
@@ -221,10 +221,89 @@ mom.std.norm <- function(j){
 
 Finally, all previously generated objects are combined to a `datMix` object.
 
-```{r faithdatmix}
+``` r
 # construct a datMix object that summarizes all the necessary information:
 faithful.dM <- datMix(faithful.obs, dist = norm.dist, discrete = norm.discrete,
                       theta.bound.list = norm.bound.list,
                       MLE.function = MLE.norm.list, Hankel.method = method,
                       Hankel.function = mom.std.norm)
 ```
+
+
+
+
+
+
+
[…]
