+++ Documentation.md
@@ -60,7 +60,7 @@ method provides more accurate estimates than the others.
 # Installation
 
 To install from `CRAN`, use:
-```
+``` r
 install.packages("mixComp")
 ```
 # Section 1. Introduction to finite mixture models and mixComp
@@ -115,7 +115,7 @@ The generation of an object of class `Mix` hinges on four central arguments: a s
 
 The following example creates two `Mix` objects, a 3-component mixture of normal distributions and a 3-component mixture of Poisson distributions. 
 
-```{r mixobj}
+``` r
 set.seed(0)
 # construct a Nix object:
 normLocMix <- Mix("norm", discrete = FALSE, w = c(0.3, 0.4, 0.3), mean = c(10, 13, 17), sd = c(1, 1, 1))
@@ -130,7 +130,8 @@ plot(poisMix, main = "3-component poisson mixture", cex.main = 0.9)
 </p>
 
 If required, random samples can be generated from these mixtures.
-```{r rmix}
+
+``` r
 # generate random samples:
 normLocRMix <- rMix(1000, obj = normLocMix)
 poisRMix <- rMix(1000, obj = poisMix)
@@ -159,14 +160,14 @@ In the table, "o" and "i" respectively stand for input used during optimization 
 
 As a simple example of a given dataset to which mixture models have been applied extensively, take the Old Faithful dataset [[29]](#29), [[1]](#1), [[18]](#18). In the context of mixture model estimation, the variable `waiting`, which gives the time in minutes between eruptions of the Old Faithful geyser in the Yellowstone National Park, is often considered to be the variable of interest. To estimate the number of components of the mixture distribution that provides a suitable approximation to the `waiting` data via **mixComp**, the raw data vector of observations has to be converted to a `datMix` object first. For the sake of exposition we specify all arguments of  the `datMix` function, starting with the vector of observations <img src="https://github.com/yuliadm/mixComp/blob/main/misc/Tex2Img_1649977791.jpg"> and the string `dist`, specifying <img src="https://github.com/yuliadm/mixComp/blob/main/misc/Tex2Img_1650014854.jpg"> and the boolean `discrete`. As has often been done in the relevant literature, we assume that the data comes from a normal mixture.
 
-```{r faithopts}
+``` r
 faithful.obs <- faithful$waiting
 norm.dist <- "norm"
 norm.discrete <- FALSE
 ```
 Second, a named list of length <img src="https://github.com/yuliadm/mixComp/blob/main/misc/Tex2Img_1649976201.jpg"> containing the bounds of <img src="https://github.com/yuliadm/mixComp/blob/main/misc/Tex2Img_1650015725.jpg"> has to be created. In this example, <img src="https://github.com/yuliadm/mixComp/blob/main/misc/Tex2Img_1650015826.jpg">.
 
-```{r normlist}
+``` r
 # define the range for parameter values:
 norm.bound.list <- list("mean" = c(-Inf, Inf), "sd" = c(0, Inf))
 ```
@@ -174,7 +175,7 @@ Next, the argument `MLE.function` contains a single function if <img src="https:
 
 Presuming a normal mixture, one specifies 2 functions, namely the MLE of the mean <img src="https://github.com/yuliadm/mixComp/blob/main/misc/Tex2Img_1650015978.jpg"> and the MLE of the standard deviation <img src="https://github.com/yuliadm/mixComp/blob/main/misc/Tex2Img_1650016036.jpg">.
 
-```{r normfun}
+``` r
 # define the MLE functions for the mean and sd: 
 MLE.norm.mean <- function(dat) mean(dat)
 MLE.norm.sd <- function(dat){
@@ -184,7 +185,7 @@ MLE.norm.list <- list("MLE.norm.mean" = MLE.norm.mean, "MLE.norm.sd" = MLE.norm.
 ```
[…]
