+++ paper/paper.md
@@ -1,188 +1,430 @@
 ---
-title: 'Estimating the Complexity of a Finite Mixture with the \proglang{R} Package \pkg{mixComp}'
+title: 'mixComp: An R package for estimating complexity of a mixture'
 tags:
   - R
-  - mixture complexity estimation
-  - mixture models
-  - Hankel matrix
-  - minimum distance
-  - likelihood-rato test
+  - mixture distribution
+  - complexity
 authors:
-  - name: Anja Weigel^[author] 
-    affiliation: 2  
-  - name: Fadoua Balabdaoui^[co-author]  
-    affiliation: 1 
-  - name: Yulia Kulagina^[co-author, maintainer] 
+  - name: Anja Weigel^[Co-first author] # note this makes a footnote saying 'Co-first author'
+    affiliation: 2 # (Multiple affiliations must be quoted)
+  - name: Fadoua Balabdaoui^[Co-first author] 
     affiliation: 1
-  - name: Lilian Mueller^[contributor]
+  - name: Yulia Kulagina^[Corresponding author] 
+    affiliation: 1
+  - name: Lilian Mueller^[Contributor]
     affiliation: 2
-  - name: Martin Maechler^[contributor]
-    (package 'nor1mix' as model, <https://orcid.org/0000-0002-8685-9910>)
-    affiliation: 1  
 affiliations:
-  - name: ETH Zurich, Seminar for Statistics, Switzerland
+ - name: ETH Zurich, Seminar for Statistics, Switzerland
    index: 1
-  - name: ETH Zurich, Switzerland
+ - name: ETH Zurich, Switzerland
    index: 2
-date: 17 April 2022
+date: 19 April 2022
 bibliography: refs.bib
 
 ---
 
 # Summary
 
-The **mixComp** package provides a number of methods for obtaining a consistent estimate of the complexity of a 
-finite mixture (the focus is made on the univariate case). The considered approaches can be loosely grouped into three categories:
-<ul>
-  <li> methods built upon the determinants of the Hankel matrix of moments of the mixing distribution; </li>
-  <li> methods based on penalized minimum distance between the unknown probability density and a consistent estimator thereof. 
-    The distances considered in this survey are the Hellinger and the $L_2$-distances; </li>
-  <li> likelihood ratio test (LRT) - based techniques. </li>
-</ul>
-While not the primary goal, most methods simultaneously estimate the component weights and parameters. 
-In this document, we give a brief overview of the methodology, and demonstrate the package's functionality in 
-both real world examples and synthetically generated data. Moreover, we show how the package can be used on virtually 
-any parametric mixture as long as functions generating random variates and evaluating the density are provided 
-for the component distribution.
+Mixture models have been used extensively in statistical applications and have attracted much attention from theoretical and computational perspectives. The list of works on mixture models is extensive: [@Teicher63], [@LindsayI], [@LindsayII], [@Titterington], [@McLachlan], to name a few. 
+
[…]
