### 2022-04-25 #4332 @danielskatz
There are also two paper.md files in your repo, one in the root and one in the paper directory.  Please remove the extra one.
