### 2022-04-25 03bdcd8
Delete paper directory
