+++ paper/paper.md
@@ -1,430 +1,0 @@
----
-title: 'mixComp: An R package for estimating complexity of a mixture'
-tags:
-  - R
-  - mixture distribution
-  - complexity
-authors:
-  - name: Anja Weigel^[Co-first author] # note this makes a footnote saying 'Co-first author'
-    affiliation: 2 # (Multiple affiliations must be quoted)
-  - name: Fadoua Balabdaoui^[Co-first author] 
-    affiliation: 1
-  - name: Yulia Kulagina^[Corresponding author] 
-    affiliation: 1
-  - name: Lilian Mueller^[Contributor]
-    affiliation: 2
-affiliations:
- - name: ETH Zurich, Seminar for Statistics, Switzerland
-   index: 1
- - name: ETH Zurich, Switzerland
-   index: 2
-date: 19 April 2022
-bibliography: refs.bib
-
----
-
-# Summary
-
-Mixture models have been used extensively in statistical applications and have attracted much attention from theoretical and computational perspectives. The list of works on mixture models is extensive: [@Teicher63], [@LindsayI], [@LindsayII], [@Titterington], [@McLachlan], to name a few. 
-
-The popularity of such models stems from the fact that they allow for modeling heterogeneous data whose distribution cannot be captured by a single parametric distribution. The (unknown) distribution is assumed to result from mixing over some latent parameter in the following sense: the latent parameter is viewed itself as a random variable drawn from some unknown mixing distribution. When this mixing distribution is only assumed to belong to the ensemble of all possible distribution functions, the mixture model is called *nonparametric* and estimation of the mixing distribution requires using some nonparametric estimation method, such as the nonparametric maximum likelihood estimator (NPMLE) whose fundamental properties were scrutinized in [@LindsayI], [@LindsayII]. One remarkable property of the NPMLE of the mixing distribution is that it is, under some conditions, a discrete distribution function with at most $k$ number of jumps, where $k$ is the number of distinct observations in the random sample drawn from the mixture. This allows for considering the smaller class of finite mixture models (characterized by a discrete mixing distribution with a finite number of jumps). In some very simple situations, the number of mixture components could be known in advance, in which case the model is *fully parametric* and convergence of classical estimators such as the parametric maximum likelihood estimator (MLE) occurs at the rate $\sqrt{n}$ (given certain conditions). Also, the well-known expectation-maximization (EM) algorithm [@Dempster] can be used to find the MLE of the unknown parameters. However, in many applications such knowledge is rarely available and the number of components has to be estimated from the data. The goal of **mixComp** is to estimate the unknown complexity using several methods. 
-
-The **mixComp** package provides several methods known from the statistical literature for estimating the unknown complexity of a (univariate) finite mixture. The considered approaches can be loosely grouped into three categories:
-
-  - methods built upon the determinants of the Hankel matrix of moments of the mixing distribution; 
-  
-  - methods based on penalized minimum distance between the unknown probability density and a consistent estimator thereof;
-   
-  - likelihood ratio test (LRT) - based techniques. 
-
[…]
