+++ paper/refs.bib
@@ -1,527 +1,0 @@
-% Encoding: UTF-8
-
-@Article{hankel,
-  author    = {Didier Dacunha-Castelle and Elisabeth Gassiat},
-  title     = {The Estimation of the Order of a Mixture Model},
-  journal   = {Bernoulli},
-  year      = {1997},
-  volume    = {3},
-  number    = {3},
-  pages     = {279--299},
-  month     = {09},
-  doi       = {10.2307/3318593},
-  fjournal  = {Bernoulli},
-  publisher = {Bernoulli Society for Mathematical Statistics and Probability},
-  url       = {https://www.jstor.org/stable/3318593},
-}
-
-@Article{l2,
-  author   = {T. Umashanger and T.N. Sriram},
-  title    = {L2E Estimation of Mixture Complexity for Count Data},
-  journal  = {Computational Statistics \& Data Analysis},
-  year     = {2009},
-  volume   = {53},
-  number   = {12},
-  pages    = {4243 - 4254},
-  issn     = {0167-9473},
-  abstract = {For count data, robust estimation of the number of mixture components in finite mixtures is revisited using L2 distance. An information criterion based on L2 distance is shown to yield an estimator, which is also shown to be strongly consistent. Monte Carlo simulations show that our estimator is competitive with other procedures in correctly determining the number of components when the data comes from Poisson mixtures. When the data comes from a negative binomial mixture but the postulated model is a Poisson mixture, simulations show that our estimator is highly competitive with the minimum Hellinger distance (MHD) estimator in terms of robustness against model misspecification. Furthermore, we illustrate the performance of our estimator for a real dataset with overdispersion and zero-inflation. Computational simplicity combined with robustness property makes the L2E approach an attractive alternative to other procedures in the literature.},
-  doi      = {10.1016/j.csda.2009.05.013},
-  url      = {http://www.sciencedirect.com/science/article/pii/S0167947309002023},
-}
-
-@Article{hell,
-  author   = {Mi-Ja Woo and T.N. Sriram},
-  title    = {Robust Estimation of Mixture Complexity for Count Data},
-  journal  = {Computational Statistics \& Data Analysis},
-  year     = {2007},
-  volume   = {51},
-  number   = {9},
-  pages    = {4379 - 4392},
-  issn     = {0167-9473},
[…]
