+++ refs.bib
@@ -242,7 +242,7 @@
   volume   = {63},
   pages    = {435--447},
   issn     = {0006-341X},
-  doi      = {10.1093/biomet/63.3.435}
+  doi      = {10.1007/978-0-387-75692-9_5}
   fjournal = {Biometrika},
   mrclass  = {62F10 (62F15 62P10)},
   mrnumber = {1925550},
