+++ refs.bib
@@ -242,7 +242,7 @@
   volume   = {63},
   pages    = {435--447},
   issn     = {0006-341X},
-  doi      = {10.1007/978-0-387-75692-9_5}
+  doi      = {10.1007/978-0-387-75692-9_5},
   fjournal = {Biometrika},
   mrclass  = {62F10 (62F15 62P10)},
   mrnumber = {1925550},
@@ -282,7 +282,7 @@
   year     = {2020},
   volume   = {204},
   pages    = {55-71},
-  doi      = {10.1016/j.jspi.2019.04.006}
+  doi      = {10.1016/j.jspi.2019.04.006},
   fjournal = {Journa of Statistical Planning and Inference},
   mrclass  = {62G07 (60G15 62E20 62G05)},
   mrnumber = {2382657},
@@ -297,7 +297,7 @@
   number    = {2},
   pages     = {255--279},
   issn      = {09528385},
-  doi       = {10.2307/2341080}
+  doi       = {10.2307/2341080},
   publisher = {[John Wiley and Sons, Royal Statistical Society]},
   url       = {http://www.jstor.org/stable/2341080},
 }
@@ -319,7 +319,7 @@
   number    = {425},
   pages     = {259--267},
   issn      = {01621459},
-  doi       = {10.1080/01621459.1994.10476467}
+  doi       = {10.1080/01621459.1994.10476467},
   abstract  = {An exact Bayesian analysis of finite mixture distributions is often computationally infeasible, because the number of terms in the posterior density grows exponentially with the sample size. A modification of the Laplace method is presented and applied to estimation of posterior functions in a Bayesian analysis of finite mixture distributions. The procedure, which involves computations similar to those required in maximum likelihood estimation, is shown to have high asymptotic accuracy for finite mixtures of certain exponential-family densities. For these mixture densities, the posterior density is also shown to be asymptotically normal. An approximation of the posterior density of the number of components is presented. The method is applied to Duncan's barley data and to a distribution of lake chemistry data for north-central Wisconsin.},
   publisher = {[American Statistical Association, Taylor and Francis, Ltd.]},
   url       = {http://www.jstor.org/stable/2291222},
@@ -334,7 +334,7 @@
   number    = {4},
   pages     = {441--453},
   issn      = {00401706},
-  doi       = {10.1080/00401706.1992.10484955}
+  doi       = {10.1080/00401706.1992.10484955},
   abstract  = {A modification of the Laplace method is presented and applied to estimation of posterior functions in a Bayesian analysis of finite-mixture distributions. The technique is nonsequential yet relatively fast and provides estimates of mixture-model parameters and classification probabilities. The method is applied to a regional distribution of lake-chemistry data for north central Wisconsin. A mixture density of two lognormal populations is estimated for the acid-neutralizing capacity of lakes in the region, using several other lake characteristics as explanatory variables for classification into lake subpopulations. The fitted mixture model provides a good representation of the observed distribution. Separation into subpopulations based solely on the other lake characteristics matches the mixture-model classification relatively well.},
   publisher = {[Taylor and Francis, Ltd., American Statistical Association, American Society for Quality]},
   url       = {http://www.jstor.org/stable/1268943},
@@ -373,7 +373,7 @@
   volume   = {24},
   number   = {2},
   pages    = {167--175},
-  doi      = {307/3315623}
+  doi      = {307/3315623},
   fjournal = {The Canadian Journal of Statistics. La Revue Canadienne de Statistique},
 }
 
@@ -398,7 +398,7 @@
   pages      = {1799–1809},
   month      = dec,
   issn       = {0167-8655},
-  doi        = {10.1016/j.patrec.2004.07.007}
+  doi        = {10.1016/j.patrec.2004.07.007},
[…]
