+++ refs.bib
@@ -222,6 +222,7 @@
   volume  = {8},
   number  = {1},
   pages   = {289--317},
+  doi     = {10.32614/rj-2016-021},
   url     = {10.32614/RJ-2016-021},
 }
 
@@ -256,6 +257,7 @@
   volume   = {93},
   pages    = {107--118},
   issn     = {0167-9473},
+  doi      = {10.1016/j.csda.2014.10.021},
   fjournal = {Computational Statistics and Data Analysis},
   mrclass  = {62G05},
   mrnumber = {3406199},
@@ -362,6 +364,7 @@
   volume  = {51},
   number  = {12},
   pages   = {1--25},
+  doi     = {10.18637/jss.v051.i12},
   url     = {http://www.jstatsoft.org/v51/i12/},
 }
 
@@ -373,7 +376,7 @@
   volume   = {24},
   number   = {2},
   pages    = {167--175},
-  doi      = {307/3315623},
+  doi      = {10.2307/3315623},
   fjournal = {The Canadian Journal of Statistics. La Revue Canadienne de Statistique},
 }
 
@@ -465,6 +468,7 @@
   number     = {1},
   pages      = {1--38},
   issn       = {0035-9246},
+  doi        = {10.1111/j.2517-6161.1977.tb01600.x},
   note       = {With discussion},
   fjournal   = {Journal of the Royal Statistical Society. Series B. Methodological},
   mrclass    = {62F10},
@@ -516,6 +520,7 @@
   mrnumber   = {838090},
   mrreviewer = {S. John},
   pages      = {x+243},
+  doi        = {10.2307/2531224},
 }
 
 @Book{McLachlan,
