+++ paper.md
@@ -187,7 +187,7 @@ As the squared $L_2$ distance might involve an infinite sum (for distributions w
 $$\frac{0.6}{n} \ln\left(\frac{j+1}{j}\right) \quad\quad \text{ and} \quad\quad \frac{0.6 \ln(n)}{n} \ln\left(\frac{j+1}{j}\right)$$
 can be used. If a customized function is used, its arguments have to be named `j` and `n`. The reader is invited to consult [@l2] for further details.
 
-#### 2. `hellinger.disc`
+#### (ii) `hellinger.disc`
 
 This function presents an alternative estimation procedure for *discrete* mixtures, working similarly to `L2.disc`. It is based on the squared Hellinger distance, causing the 'best' estimate $(\hat{\mathbf{w}}^j, \hat{\mathbf{\theta}}^j) \in W_j \times \Theta_j$ for a given $j$ to equal
 $$(\hat{\mathbf{w}}^j, \hat{\mathbf{\theta}}^j) = \text{argmin}_{(\mathbf{w}, \mathbf{\theta})} H^2(f_{j, \mathbf{w}, \mathbf{\theta}}, \tilde{f_n}) 
@@ -196,7 +196,7 @@ with $X_{(n)} = \max_{i = 1}^n (X_i)$. The relevant theory can be found in [@hel
 $$\text{AIC} = \frac{d+1}{n} \quad \quad \text{and} \quad \quad \text{SBC} = \frac{(d+1)\ln(n)}{2n}$$
 (where $d$ is the dimension, i.e., $\Theta \subseteq \mathbb{R}^d$).  
 
-#### 3. `hellinger.cont`
+#### (iii) `hellinger.cont`
  
 This procedure is applicable to *continuous* mixture models and uses a kernel density estimator (KDE) as $\tilde{f}_n$. Its `bandwidth` can be chosen by the user, or the adaptive KDE found in [@adap, p. 1720, equation 2] may be used by specifying `bandwidth = "adaptive"`. The calculations are based on the continuous version of the squared Hellinger distance, where the 'best' estimate $(\hat{\mathbf{w}}^j, \hat{\mathbf{\theta}}^j) \in W_j \times \Theta_j$ for a given $j$ corresponds to
 \begin{equation}\label{eq:hellcont}
@@ -230,7 +230,7 @@ Nonetheless, the user will get a warning when using one of non-consistent predef
 
 The preceding example shows that $\hat{p}$ depends on the chosen threshold $t(j, n)$, the choice of which always remains somewhat arbitrary. It is thus desirable to have versions of the preceding functions which do not suffer from this drawback. `L2.boot.disc`, `hellinger.boot.disc` and `hellinger.boot.cont` all work similarly to their counterparts, with the exception that the difference in distances is not compared to a predefined threshold but to a value generated by a bootstrap procedure.  At every iteration, the procedure sequentially tests $p = j$ versus $p = j+1$ for $j = 1,2, \dots$, using a parametric bootstrap to generate `B` samples of size $n$ from a $j$-component mixture given the previously calculated 'best' parameter values $(\hat{\mathbf{w}}^j, \hat{\mathbf{\theta}}^j)$. For each of the bootstrap samples, the 'best' estimates corresponding to densities with $j$ and $j+1$ components are calculated, as well as their difference in distances from $\tilde{f}_n$. The null hypothesis $H_0: p = j$ is rejected and $j$ is increased by $1$ if the original difference $D(\hat{f}_j, \tilde{f}_n) - D(\hat{f}_{j+1}, \tilde{f}_n)$ lies outside of the interval $[ql, qu]$, specified by the `ql` and `qu` empirical quantiles of the bootstrapped differences. Otherwise, $j$ is returned as the order estimate $\hat{p}$. 
 
[…]
