### 2022-04-25 #4332 @yuliadm
@danielskatz Dear Daniel, thanks a lot for your help! I think everything should be OK now! I have fixed all DOI's, and the .pdf compiles now. I've also removed the paper folder from the repository.
