### 2022-04-25 #4332 @danielskatz
@yuliadm - this paper is now almost 5000 words long, while the JOSS guidelines are 1000.  I'm not sure what happened, since you said you had shortened the paper before, but you need to shorten it again.
