### 2022-04-25 #4332 @jbytecode
Dear @yuliadm,

It seems the paper is over-sized and you are trying to shrink the words count. Here is a list of some suggestions:

- The summary is too long, it is a little bit like a introduction, it can be shortened
- The examples are too verbose, code outputs can be omitted.
- General framework and the functions sections may be combined to form a more compact section.
