+++ paper.md
@@ -25,7 +25,7 @@ bibliography: refs.bib
 
 # Summary
 
-Mixture models have been used broadly in statistical applications and studied extensively ([@Teicher63], [@LindsayI], [@LindsayII], [@Titterington], [@McLachlan], and other works). They allow for modeling heterogeneous data whose distribution cannot be captured by a single parametric distribution. The (unknown) distribution is assumed to result from mixing over some latent parameter in the following sense: the latent parameter is viewed as a random variable drawn from some unknown mixing distribution. When this mixing distribution is only assumed to belong to the ensemble of all possible distribution functions, the mixture model is called *nonparametric* and estimation of the mixing distribution requires using some nonparametric estimation method, such as the nonparametric maximum likelihood estimator (NPMLE) (see [@LindsayI], [@LindsayII]). One remarkable property of the NPMLE of the mixing distribution is that, it is, under some conditions, a discrete distribution with at most $k$ number of jumps, where $k$ is the number of distinct observations in the random sample drawn from the mixture. This allows for considering a smaller class of finite mixture models (characterized by a discrete mixing distribution with a finite number of jumps). In simple situations, the number of mixture components could be known in advance, in which case the model is *fully parametric* and convergence of classical estimators such as the parametric maximum likelihood estimator (MLE) occurs at the rate $\sqrt{n}$ (given certain conditions). Also, the well-known expectation-maximization (EM) algorithm ([@Dempster]) can be used to find the MLE of the unknown parameters. However, in many applications such knowledge is rarely available and the number of components has to be estimated from the data. The goal of **mixComp** is to estimate the unknown complexity using several methods. 
+Mixture models have been used broadly in statistical applications and studied extensively [@Teicher63, @LindsayI, @LindsayII, @Titterington, @McLachlan, etc]. They allow for modeling heterogeneous data whose distribution cannot be captured by a single parametric distribution. The (unknown) distribution is assumed to result from mixing over some latent parameter in the following sense: the latent parameter is viewed as a random variable drawn from some unknown mixing distribution. When this mixing distribution is only assumed to belong to the ensemble of all possible distribution functions, the mixture model is called *nonparametric* and estimation of the mixing distribution requires using some nonparametric estimation method, such as the nonparametric maximum likelihood estimator (NPMLE) [@LindsayI], [@LindsayII]. One remarkable property of the NPMLE of the mixing distribution is that, it is, under some conditions, a discrete distribution with at most $k$ number of jumps, where $k$ is the number of distinct observations in the random sample drawn from the mixture. This allows for considering a smaller class of finite mixture models (characterized by a discrete mixing distribution with a finite number of jumps). In simple situations, the number of mixture components could be known in advance, in which case the model is *fully parametric* and convergence of classical estimators such as the parametric maximum likelihood estimator (MLE) occurs at the rate $\sqrt{n}$ (given certain conditions). Also, the well-known expectation-maximization (EM) algorithm [@Dempster] can be used to find the MLE of the unknown parameters. However, in many applications such knowledge is rarely available and the number of components has to be estimated from the data. The goal of **mixComp** is to estimate the unknown complexity using several methods. 
 
 The **mixComp** package provides several methods for estimating the unknown complexity of a (univariate) finite mixture that can be arranged into 3 categories:
 
@@ -63,7 +63,7 @@ Assume that the family of the component densities $\{g(x; \theta):\theta \in \Th
 
 # 1. Functions using Hankel matrices
 
-The basic approach (see [@hankel]) estimates the mixture order as
[…]
