### 2022-04-25 #4332 @jbytecode
Dear @yuliadm

I really appreciate your effort. Thank you.

Yet another suggestion is to remove some theoretical information and definition and to focus on the software itself.
