### 2022-04-25 #4332 @yuliadm
Dear @jbytecode thanks a lot for the suggestion! I am trying to leave as little theory as possible, I'll think about removing the definition, although I'd like to keep a little bit of math:)
