### 2022-04-26 #4332 @jbytecode
The current word count is still %11 larger than the pre-defined upper limit, I think It is beyond my authority to make this decision.
