### 2022-04-26 #4332 @jbytecode
👋👋👋 Dear @zhiiiyang  👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**mixComp: An R package for estimating complexity of a mixture**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/4332).
