### 2022-04-27 #4354 @editorialbot
[REVIEW]: mixComp: An R package for estimating complexity of a mixture
**Submitting author:** @yuliadm (<a href="http://orcid.org/0000-0002-9066-7129">Yulia</a>)
**Repository:** https://github.com/yuliadm/mixComp
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.3
