### 2022-04-28 yuliadm/mixComp#1 @zhiiiyang
Please add unit tests
Hi, I am the reviewer from JOSS (https://github.com/openjournals/joss-reviews/issues/4354). I don't see any unit tests available. Do you mind adding at least one unit test for each main function? Thank you!
