+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(mixComp)
+library(testthat)
+
+test_check("mixComp")
