### 2022-04-30 yuliadm/mixComp#1 @yuliadm
Dear @zhiiiyang  I've added the tests. Hope it's ok now, otherwise let me know! Thanks!
