### 2022-05-01 yuliadm/mixComp#1 @zhiiiyang
Thank you for adding all the unit tests. One of them failed. Could you resolve it?
`devtools::test()`

```
✓ |         4 | mix.lrt [113.2s]
✓ |         5 | Mix
x | 1       5 | nonparamHankel [0.3s]
────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
Failure (test-nonparamHankel.R:84:3): function checks
round(res[2], 3) not equal to 0.002.
1/1 mismatches
[1] 0.009 - 0.002 == 0.007
Duration: 602.1 s

[ FAIL 1 | WARN 0 | SKIP 0 | PASS 76 ]
```
