### 2022-05-01 yuliadm/mixComp#2 @zhiiiyang
Miscellaneous requests
Hi, I am the reviewer from JOSS (https://github.com/openjournals/joss-reviews/issues/4354). could you add instructions on how to install this package from GitHub besides from CRAN? Here is an example from ggplot2. Also, could you add a branch called dev for changes? The dev branch can be merged into the main branch once reaching to a stable version.

```
# The easiest way to get ggplot2 is to install the whole tidyverse:
install.packages("tidyverse")

# Alternatively, install just ggplot2:
install.packages("ggplot2")

# Or the development version from GitHub:
# install.packages("devtools")
devtools::install_github("tidyverse/ggplot2")
```
