### 2022-05-06 yuliadm/mixComp#2 @zhiiiyang
@yuliadm, thank you. I also have other miscellaneous suggestions. I will summarize it below to avoid opening too many issues for only 1~2 requests at once.

- [x] add instructions on installing it from GitHub
- [x] move reproduction_script.R and reproduction_script.html to the vignettes folder, which is empty right now.
- [x] create a paper folder and move paper.md and refs.bib to the folder
- [x] is figures folder the same as images folder? what is this misc folder?
- [x] add Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
