+++ Documentation.md
@@ -140,8 +140,8 @@ plot(normLocRMix, main = "Three component normal mixture", cex.main = 0.9)
 plot(poisRMix, main = "Three component poisson mixture", cex.main = 0.9)
 ```
 <p float="left">
-<img src="https://github.com/yuliadm/mixComp/blob/main/figures/normRMix.png" />
-<img src="https://github.com/yuliadm/mixComp/blob/main/figures/poisRMix.png" />
+<img src="https://github.com/yuliadm/mixComp/blob/main/figures/normRMix.png" width="500" height="400"/>
+<img src="https://github.com/yuliadm/mixComp/blob/main/figures/poisRMix.png" width="500" height="400"/>
 </p>
   
 The third object class shown in Table 1, called `datMix`, represents the data vector <img src="https://github.com/yuliadm/mixComp/blob/main/misc/Tex2Img_1649977791.jpg"> based on which the mixture complexity is supposed to be estimated. These objects are most central to the package, as every procedure estimating the order of a mixture takes a `datMix` object as input. Apart from <img src="https://github.com/yuliadm/mixComp/blob/main/misc/Tex2Img_1649977791.jpg">, it contains other "static" information needed for the estimation procedure (in contrast to "tuning parameters", which can be changed with every function call. An example of such a tuning parameter is the number of bootstrap replicates for a function employing a bootstrap procedure). A brief overview of which "static" attributes need to be supplied for each complexity estimation routine is given in Table 2. 
@@ -383,8 +383,8 @@ plot(normdets_sca_pen, main = "3-component Normal mixture", cex.main = 0.9)
 ```
 
 <p float="left">
-<img src="https://github.com/yuliadm/mixComp/blob/main/images/np_art_1.png" />
-<img src="https://github.com/yuliadm/mixComp/blob/main/images/np_art_2.png" />
+<img src="https://github.com/yuliadm/mixComp/blob/main/figures/np_art_1.png" width="500" height="400"/>
+<img src="https://github.com/yuliadm/mixComp/blob/main/figures/np_art_2.png" width="500" height="400"/>
 </p>
 
 Having created the data ourselves, we know that it comes from a 3-component Poisson mixture and a 3-component Gaussian mixture respectively. The resulting plots indicate that while theoretically sound, the scaled version of the Hankel method can struggle to correctly identify the number of components in practice.
@@ -407,8 +407,8 @@ plot(norm_sca_pen)
 ```
 
 <p float="left">
-<img src="https://github.com/yuliadm/mixComp/blob/main/images/p_art_1.png" />
-<img src="https://github.com/yuliadm/mixComp/blob/main/images/p_art_2.png" />
+<img src="https://github.com/yuliadm/mixComp/blob/main/figures/p_art_1.png" width="500" height="400"/>
+<img src="https://github.com/yuliadm/mixComp/blob/main/figures/p_art_2.png" width="500" height="400"/>
 </p>
 
[…]
