### 2022-05-06 yuliadm/mixComp#2 @yuliadm
Hi @zhiiiyang Many thanks for your optimization suggestions! I've found them great and have implemented most of them (in the dev branch):

-  added the instructions to the .md file
- moved the reproduction files to the vignette folder
-  moved the paper files to the newly created paper folder
- merged the figures and images folder into one.

I kept the misc folder as it contains formulae for the Documentation.md file (this contains the detailed manual for those who want to have it directly on Git without going to the vignettes).
