### 2022-05-07 yuliadm/mixComp#2 @zhiiiyang
@yuliadm, thank you for making the changes. I have checked the to-do list here (https://github.com/yuliadm/mixComp/issues/2#issuecomment-1119274904). One more request, could you add community guidelines for contributions in your readme.md? Once you are done, I think you can merge the current dev branch with the main branch to incorporate those changes. You may want to bump the software version.
