### 2022-05-08 yuliadm/mixComp#3 @zhiiiyang
Regarding the progress output 
Hi, I am the reviewer from JOSS (https://github.com/openjournals/joss-reviews/issues/4354). I just ran the Shakespeare example from the paper. I wonder if you can reduce the progress output in order to make it concise. For example, there is no need to print a message for every iteration. Instead, you can print one message for every estimation (since it runs fast anyway) or display a progress bar (https://medium.com/human-in-a-machine-world/displaying-progress-in-a-loop-or-function-r-664796782c24).

![Screen Shot 2022-05-07 at 10 55 32 PM](https://user-images.githubusercontent.com/7269113/167283828-fe311179-a639-4df7-a68a-75ce6bb3bdf1.png)
