### 2022-05-09 #4354 @zhiiiyang
@jbytecode @yuliadm I have completed my review. Please see all the comments above. https://github.com/openjournals/joss-reviews/issues/4354#issuecomment-1111224747

@yuliadm has completed some requests in the dev branch and will merge them into the main branch once all requests in https://github.com/yuliadm/mixComp/issues/2 are addressed. @Athene-ai, you might want to review the dev branch for the latest commits if the merge hasn't been completed by that time.
