### 2022-05-09 #4354 @zhiiiyang
@jbytecode, I see. I just check off the checklist and posted all my comments below in a separate post.

### Additional comments
- Please add the unit tests https://github.com/yuliadm/mixComp/issues/1
- Please reorganize the files into the corresponding folders. See more details https://github.com/yuliadm/mixComp/issues/2
- Please add community guidelines https://github.com/yuliadm/mixComp/issues/2
- Since the R package only has commits from Yulia. May I ask for the contributions of the other three co-authors?
- Can authors comment on another R package, rebmix, compared to mixComp?
- Please reduce unnecessary progress output to simplify printed output. https://github.com/yuliadm/mixComp/issues/3
- For someone who doesn't know the `MLE.function` for certain, is there a convenient way to compare across different distributions and select the optimal one？
- In readme.md, it said that `nonparamHankel` might not produce correct results. In https://github.com/yuliadm/mixComp#the-children-dataset, it used both functions `paramHankel` and `paramHankel.scaled` to conclude that the cluster is equal to 2. If we can't make results without having consistent results from both functions, should it be set as default? What if these two functions return different results?
- Is the plot over different complexity `j` only available for `nonparamHankel`? Can we have that implemented for results from `hellinger.cont` or `hellinger.disc`？
