### 2022-05-09 yuliadm/mixComp#2 @yuliadm
Hi @zhiiiyang ! I have added the Community Guidelines to the readme file.
