### 2022-05-20 yuliadm/mixComp#4 @Athene-ai
Package installation procedure
Hi, I am the reviewer from JOSS ([#4354](https://github.com/openjournals/joss-reviews/issues/4354)).

I suggest adding other information about the installation procedure also for not R expert
