### 2022-05-20 yuliadm/mixComp#5 @Athene-ai
Provide some real example
Hi, I am the reviewer from JOSS ([#4354](https://github.com/openjournals/joss-reviews/issues/4354)).

Is it possible to add some examples from real data?
