### 2022-05-20 yuliadm/mixComp#6 @Athene-ai
Different type of Mix package
Hi, I am the reviewer from JOSS (https://github.com/openjournals/joss-reviews/issues/4354).

Is it possible for the author provides a better explanation of the different type of Mix package and create a demonstration on how they works
