### 2022-05-20 #4354 @Athene-ai
@jbytecode @yuliadm I have completed my review [#4354 (comment)](https://github.com/openjournals/joss-reviews/issues/4354#issuecomment-1133031070)  also adding further comments below

**Additional comments**

-  Please revise the installation procedure (see [Package installation procedure](https://github.com/yuliadm/mixComp/issues/4#issue-1243336630))
- Please see this about a real data experiment (see [Provide some real example](https://github.com/yuliadm/mixComp/issues/5#issue-1243341270)).
-  Please better specify the existing difference between different Mix packages (see [Different type of Mix package](https://github.com/yuliadm/mixComp/issues/6#issue-1243348463)).
