### 2022-05-21 yuliadm/mixComp#4 @yuliadm
@Athene-ai Hi! There are 2 options of the package installation: a standard R procedure with install.packages and instructions on installing from GitHub using devtools.
I myself do not know any more... If you have something particular in mind, could you please share it?
