### 2022-05-21 yuliadm/mixComp#6 @yuliadm
Hi @Athene-ai ! I am a bit confused about your suggestion. What is exactly meant by the different types of Mix package? Is it about the different types of functions included in the package? They are explained... The documentation.md file, the readme.md file as well as the vignettes and manuals have a very detailed explanation of each function with multiple examples on their usage... I only included the vital information  (the methodology)  to the paper.md file due to the word count restrictions. I am just not sure which of the files you mean...
