### 2022-05-21 yuliadm/mixComp#5 @yuliadm
Hi @Athene-ai. There are several examples that are using the real data (the Old faithful data, Children data, Shakespeare data and the Acidity data) which can be found in the documentation.md as well as in the vignettes. Where should I incorporate these examples? The paper.md file? I could do that if adding the R code will not affect the word count (I am not sure whether it does)... Thanks!
