### 2022-05-21 yuliadm/mixComp#5 @Athene-ai
@yuliadm I suggest to add them in paper.md as a comment
