### 2022-05-21 yuliadm/mixComp#4 @Athene-ai
Hi @yuliadm I tried both methods but install.package one does not work
