### 2022-05-21 yuliadm/mixComp#6 @Athene-ai
Hi @yuliadm I mean difference between  other packages, like rebmix, compared with your and why user should use your
