+++ paper/refs.bib
@@ -537,3 +537,9 @@
   pages      = {xxii+419},
 }
 
+@article{rebmix,
+  title={Multivariate normal mixture modeling, clustering and classification with the rebmix package},
+  author={Nagode, Marko},
+  journal={arXiv preprint arXiv:1801.08788},
+  year={2018}
+}
