### 2022-05-24 yuliadm/mixComp#6 @yuliadm
Hello @Athene-ai, this part is done!
