### 2022-06-01 acdb109
Update 5_lrt_utils.R
