+++ R/5_lrt_utils.R
@@ -94,6 +94,10 @@ mix.lrt <- function(obj, j.max = 10, B = 100, quantile = 0.95, control = c(trace
     j0 <- j0 + 1 # current complexity estimate
     j1 <- j0 + 1
 
+    #####################################
+    message(paste("Testing for ", j0, " components.\n", sep = ""))
+    #####################################
+    
     if(j0 > 1){ # if j1 was calculated in the last interation, pass it over to j0...
 
       mle.est0 <- mle.est1
@@ -182,9 +186,12 @@ mix.lrt <- function(obj, j.max = 10, B = 100, quantile = 0.95, control = c(trace
 
         # don't include first iteration as this just uses the original data
         # to calculate t0
-        message(paste("Running bootstrap iteration ", bs_iter, " testing for ", j0,
-                  " components.\n", sep = ""))
-
+        #message(paste("Running bootstrap iteration ", bs_iter, " testing for ", j0, " components.\n", sep = ""))
+        Sys.sleep(0.05)
+        progress(bs_iter)
+        Sys.sleep(0.05)
+        if (bs_iter==B) message(paste("Done! \n"))
+        
       } else message(paste("\n"))
 
       # in the bootstrap we have to calculate the values for j0 and j1 as the bootstrap
