### 2022-06-01 70ba9f4
Update 4_distances_cont.R
