+++ R/4_distances_cont.R
@@ -473,6 +473,10 @@ hellinger.boot.cont <- function(obj, bandwidth, j.max = 10, B = 100, ql = 0.025,
     j0 <- j0 + 1 # current complexity estimate
     j1 <- j0 + 1
 
+    #####################################
+    message(paste("Testing for ", j0, " components.\n", sep = ""))
+    #####################################
+    
     if(j0 != 1){ # pass on constraints for bootstrap
 
       ineq.j0 <- ineq.j1
@@ -617,8 +621,11 @@ hellinger.boot.cont <- function(obj, bandwidth, j.max = 10, B = 100, ql = 0.025,
 
         # don't include first iteration as this just uses the original data
         # to calculate t0
-        message(paste("Running bootstrap iteration ", bs_iter, " testing for ", j0,
-                  " components.\n", sep = ""))
+        #message(paste("Running bootstrap iteration ", bs_iter, " testing for ", j0, " components.\n", sep = ""))
+        Sys.sleep(0.05)
+        progress(bs_iter)#, progress.bar = TRUE)
+        Sys.sleep(0.05)
+        if (bs_iter==B) message(paste("Done! \n"))
 
       } else message(paste("\n"))
 
