### 2022-06-01 5efa2df
Update 3_distances_disc.R
