+++ R/3_distances_disc.R
@@ -289,6 +289,10 @@ L2.boot.disc <- function(obj, j.max = 10, n.inf = 1000, B = 100, ql = 0.025, qu 
     j0 <- j0 + 1 # current complexity estimate
     j1 <- j0 + 1
 
+    #####################################
+    message(paste("Testing for ", j0, " components.\n", sep = ""))
+    #####################################
+    
     if(j0 > 1){ # if j1 was calculated in the last interation, pass it over to j0...
 
       theta.j0 <- theta.j1
@@ -367,9 +371,12 @@ L2.boot.disc <- function(obj, j.max = 10, n.inf = 1000, B = 100, ql = 0.025, qu 
 
         # don't include first iteration as this just uses the original data
         # to calculate t0
-        message(paste("Running bootstrap iteration ", bs_iter, " testing for ", j0,
-                  " components.\n", sep = ""))
-
+        #message(paste("Running bootstrap iteration ", bs_iter, " testing for ", j0, " components.\n", sep = ""))
+        Sys.sleep(0.05)
+        progress(bs_iter)
+        Sys.sleep(0.05)
+        if (bs_iter==B) message(paste("Done! \n"))
+        
       } else message(paste("\n"))
 
       # in the bootstrap we have to calculate the values for j0 and j1 as the bootstrap
