+++ paper/paper.md
@@ -46,10 +46,8 @@ Two main features distinguish it from other mixture-related **R** [@R] packages:
 
 - **mixComp** is applicable to parametric mixtures beyond those whose component distributions are included in the **stats** package, making it more customizable than most packages for model-based clustering. 
 
-Other packages dealing with mixture models are **mclust** [@mclust], which fits Gaussian mixtures using the EM algorithm, **MixSim** [@mixsim], which allows for simulating from mixtures and comparing the performance of clustering algorithms, and **mixdist** [@mixdist], used for grouped conditional data. Aside from **mixtools**'s [@mixtools] focus on mixture-of-regressions and non-parametric mixtures, it is used to fit (multivariate) normal, multinomial or gamma mixtures with the EM algorithm, also containing routines for selecting the number of components based on information criteria and parametric bootstrapping of the LRT statistic values. However, they are limited to multinomial, normal mixtures and mixtures-of-regressions. **flexmix** [@flexmix1; @flexmix2; @flexmix3] handles mixtures-of-regression and stands out due to its extensibility, a design principle that we also aimed for.
-**rebmix**[@rebmix], dealing with univariate and multivariate finite mixture model for generation, estimation, clustering, classification purposes, offers a wide variety of well-recognized methods for mixture model estimation for both discrete and continuous variables. The approaches suggested in **mixComp** are however not among those used in **rebmix**, thus complementing it rather than providing competition. 
-
-**mixComp** can be used on virtually any parametric mixture as long as functions generating random variates and evaluating the density are provided for the component distributions.
+Other packages dealing with mixture models are **mclust** [@mclust], which fits Gaussian mixtures using the EM algorithm, **MixSim** [@mixsim], which allows for simulating from mixtures and comparing the performance of clustering algorithms, and **mixdist** [@mixdist], used for grouped conditional data. **mixtools** [@mixtools] focuses on mixture-of-regressions and non-parametric mixtures and is used to fit (multivariate) normal, multinomial or gamma mixtures with the EM algorithm, also containing routines for selecting the number of components based on information criteria and parametric bootstrapping of the LRT statistic values. However, they are limited to multinomial, normal mixtures and mixtures-of-regressions. **flexmix** [@flexmix1; @flexmix2; @flexmix3] handles mixtures-of-regression and stands out due to its extensibility, a design principle that we also aimed for.
+**rebmix**[@rebmix], dealing with univariate and multivariate finite mixture model for generation, estimation, clustering, classification purposes, offers a wide variety of recognized methods for mixture model estimation for both discrete and continuous variables. The approaches suggested in **mixComp** are however not among those used in **rebmix**, thus complementing it rather than providing competition. 
 
 # Methods
 
@@ -83,7 +81,7 @@ and the non-parametric density/mass estimate $\tilde{f}_n(x)$. Then
 $$\hat{p} = \min_j \big\{D(\hat{f}_j, \tilde{f}_n) - D(\hat{f}_{j+1}, \tilde{f}_n) \leq t(j,n) \big\},$$ 
 where $D$ denotes the distance measure, $t(j,n)$ - a suitable penalty function.
 
-**mixComp** offers several distance-based procedures described in [@l2; @hell; @hellcont]. 
+**mixComp** offers several distance-based procedures following [@l2; @hell; @hellcont]. 
 
 ### 3. Functions using LRTS
 
