### 2022-06-01 yuliadm/mixComp#5 @yuliadm
Dear @Athene-ai  I have added another real data example to the paper. At the moment it is not included as a comment, but hopefully it is fine.
