### 2022-06-01 yuliadm/mixComp#4 @yuliadm
Dear @Athene-ai I have also specifically checked both methods, and everything works for me. Could you please check again? Please note, that for installing packages using the standard R proc the command is install.packages(), and not install.package() even if one is installing only one package.
