### 2022-06-03 #4354 @yuliadm
Dear @zhiiiyang
Thank you so much for your review, your questions and comments!
I will do my best to answer them in as much detail as I can in this post.

_>Since the R package only has commits from Yulia. May I ask for the contributions of the other three co-authors?_

Handling GitHub repo is exactly my part of work as per agreement with my collaborators.
As for the contributions, Lilian Mueller has written most of the functions, Anja Weigel helped to review and optimize them and wrap them up into the package. Fadoua Balabdaoui was providing ideas, and supervising the project (including the theoretical part).

_>Can authors comment on another R package, rebmix, compared to mixComp?_

I have added the  comments on this package to the paper.md file.

_>Please reduce unnecessary progress output to simplify printed output._

I've replaced it with the progress bar where possible.

_>For someone who doesn't know the MLE.function for certain, is there a convenient way to compare across different distributions and select the optimal one?_

Thanks for this comment. Computing the maximum likelihood estimator (which maximizes the likelihood function over the parameter space) is of very common practice as it can be justified with statistical theory of the well-known properties of the Kullback-Leibler diversion. It gives a systematic way of estimating the unknown densities by finding the most likely parameters given the observed data.   In this sense, the MLE gives the optimal density given the observed sample.  Note that the MLE is so widely used that we are confident that the readers of JOSS will be able to understand the main idea. In case we need to add explanations to this regard, we will be very happy to do so.

_>In readme.md, it said that nonparamHankel might not produce correct results. In https://github.com/yuliadm/mixComp#the-children-dataset, it used both functions paramHankel and paramHankel.scaled to conclude that the cluster is equal to 2. If we can't make results without having consistent results from both functions, should it be set as default? What if these two functions return different results?_

Thanks for this question! I think there is some misunderstanding here however: the Children data is the real dataset, which means basically that there is no way to verify whether the number of components identified by a certain method is right or wrong (but even in this case the two methods in readme.md (nonparamHankel and paramHankel.scaled) agree on 2 components, which is also in line with other methods found in the literature).

Now, it is possible to say for the simulated data that the number of components is identified incorrectly (which is the case for the Poisson mixture with w = c(0.45, 0.45, 0.1) and lambda = c(1, 5, 10)). This estimation is however doe for a specific seed, and for another seed (a different sample) a different result may be obtained. In order to estimate the statistical accuracy of the method this experiment should be run a number of times (e.g. 100), then one could compute the number of times the method hits the right complexity. We are looking here at just one example, and it is perfectly normal that in this case the method underestimates the number of components: none of the methods are 100% accurate...

The last comment concerns the consistency. It is specifically stated in the paper that all of the methods considered come with the guarantees for consistency (the proofs can be found in the literature to which we refer in the relevant sections).

_>Is the plot over different complexity j only available for nonparamHankel? Can we have that implemented for results from hellinger.cont or hellinger.disc?_

Many thanks for this interesting question.  Such a plot was not  produced for the following reasons:  (1)  The plot of the penalized Hankel determinant versus the assumed complexity/order of the mixture is the only visual tool that one has for this method because estimation of the parameters of the mixture is completely avoided. Thus, the very purpose of the method is best described by such a plot. (2)  With the minimum distance methods (and also Likelihood ratio based method), it is not a common practice to produce such a plot. Indeed, the procedure is sequential in the sense that we start with one component ( complexity assumed at this stage to be equal to 1)  and then increase this number until there is only a  marginal improvement of the chosen distance.  Thus, the plot in this case would be rather non-informative as the selected complexity is the one obtained when the algorithm stops and one knows exactly that the plot should be decreasing. On the other hand, one can plot the obtained estimated density or probability mass when the selection algorithm stops. This is not at all possible with the Hankel matrix method which is only concerned with estimating the true complexity.

I very much hope this resolves most of the issues!
