### 2022-06-03 #4354 @yuliadm
Dear @Athene-ai

Thank you for your comments and suggestions. I have also tried to incorporate them:

_>Please revise the installation procedure (see https://github.com/yuliadm/mixComp/issues/4#issue-1243336630)_
resolved

_>Please see this about a real data experiment (see https://github.com/yuliadm/mixComp/issues/5#issue-1243341270)_
I've added the relevant information to the paper.md file

_>Please better specify the existing difference between different Mix packages (see https://github.com/yuliadm/mixComp/issues/6#issue-1243348463)_

There was one real data example in the paper.md already (the Shaespeare dataset), but I added one more - the Acidity data (the continuous case).

I hope I didn't misunderstand anything.
