### 2022-06-04 yuliadm/mixComp#3 @zhiiiyang
@yuliadm, could you point me to the code where you made changes to the printed message? I still got a lot of progress output after re-installing the package.
