### 2022-06-04 #4354 @zhiiiyang
@yuliadm, thank you so much for addressing my comments. I don't have any further questions except for following up with the requested change on the progress output below.

1. >Please reduce unnecessary progress output to simplify printed output.
I've replaced it with the progress bar where possible.
- Could you point me to the changed code? I still got a lot of progress output when running the example in the paper.md.
