### 2022-06-04 #4354 @yuliadm
Dear @zhiiiyang, sure!

I have changed the following files:
mixComp/R/5_lrt_utils.R
lines: 187-193
mixComp/R/4_distances_cont.R
lines: 622-628
mixComp/R/3_distances_disc.R
lines: 372-378
