### 2022-06-04 #4354 @zhiiiyang
@yuliadm, could you also change the codes for function `hellinger.boot.disc`?https://github.com/yuliadm/mixComp/blob/51aa8117ac67b0b551408e6f0722ae59d0677306/R/3_distances_disc.R#L809-L810
![Screen Shot 2022-06-04 at 12 38 31 PM](https://user-images.githubusercontent.com/7269113/172023158-16493ddd-9cb5-4230-a604-4eea316aa27e.png)
