### 2022-06-04 c7d27ec
Update 3_distances_disc.R
