### 2022-06-04 #4354 @zhiiiyang
@yuliadm, thank you for the update. I think you need to import `progress` function.
![Screen Shot 2022-06-04 at 4 32 51 PM](https://user-images.githubusercontent.com/7269113/172028808-edc86252-df04-4925-8ab1-9678587a14d0.png)

Could you test this example in your paper and make sure it is error-free after you make the updates? Also, don't forget to rerun `devtool::test()` to ensure all the unit tests pass. Thank you!
```
library(mixComp)
shakespeare.obs <- unlist(shakespeare) - 1
# define the MLE function:
MLE.geom <- function(dat) 1 / (mean(dat) + 1)
# create the datMix object:
Shakespeare.dM <- datMix(shakespeare.obs, dist = "geom", discrete = TRUE,
                         MLE.function = MLE.geom, theta.bound.list = list(prob = c(0, 1)))
# estimate the complexity:
set.seed(0)
(res <- hellinger.boot.disc(Shakespeare.dM, B = 50, ql = 0.025, qu = 0.975))
```
