### 2022-06-10 94d3d6c
Update 3_distances_disc.R
