### 2022-06-13 cf47741
Update 3_distances_disc.R
