### 2022-06-13 cea1fae
Update 3_distances_disc.R
