+++ R/4_distances_cont.R
@@ -622,10 +622,9 @@ hellinger.boot.cont <- function(obj, bandwidth, j.max = 10, B = 100, ql = 0.025,
         # don't include first iteration as this just uses the original data
         # to calculate t0
         #message(paste("Running bootstrap iteration ", bs_iter, " testing for ", j0, " components.\n", sep = ""))
-        Sys.sleep(0.05)
-        progress(bs_iter)#, progress.bar = TRUE)
-        Sys.sleep(0.05)
-        if (bs_iter==B) message(paste("Done! \n"))
+        Sys.sleep(0.5)
+        progress(bs_iter)
+        #if (bs_iter==B) message(paste("Done! \n"))
 
       } else message(paste("\n"))
 
