### 2022-06-13 b7b7b83
Update 5_lrt_utils.R
