### 2022-06-14 7c176cf
Update 3_distances_disc.R
