+++ R/3_distances_disc.R
@@ -370,9 +370,9 @@ L2.boot.disc <- function(obj, j.max = 10, n.inf = 1000, B = 100, ql = 0.025, qu 
 
         # don't include first iteration as this just uses the original data
         # to calculate t0
-        #message(paste("Running bootstrap iteration ", bs_iter, " testing for ", j0, " components.\n", sep = ""))
+        # message(paste("Running bootstrap iteration ", bs_iter, " testing for ", j0, " components.\n", sep = ""))
         Sys.sleep(0.5)
-        progress(bs_iter)
+        progress(bs_iter, progress.bar = TRUE)
         #if (bs_iter==B) message(paste("Done! \n"))
         } else message(paste("\n"))
 
@@ -810,7 +810,7 @@ hellinger.boot.disc <- function(obj, j.max = 10, B = 100, ql = 0.025, qu = 0.975
         # message(paste("Running bootstrap iteration ", bs_iter, " testing for ", j0, " components.\n", sep = ""))
         
         Sys.sleep(0.5)
-        progress(bs_iter)
+        progress(bs_iter, progress.bar = TRUE)
         #if (bs_iter==B) message(paste("Done! \n"))
         
       } else message(paste("\n"))
