### 2022-06-14 0f3ecb3
Update 3_distances_disc.R
