+++ R/3_distances_disc.R
@@ -371,9 +371,9 @@ L2.boot.disc <- function(obj, j.max = 10, n.inf = 1000, B = 100, ql = 0.025, qu 
         # don't include first iteration as this just uses the original data
         # to calculate t0
         # message(paste("Running bootstrap iteration ", bs_iter, " testing for ", j0, " components.\n", sep = ""))
-        Sys.sleep(0.5)
-        progress(bs_iter, progress.bar = TRUE)
-        #if (bs_iter==B) message(paste("Done! \n"))
+          progress(bs_iter, progress.bar = TRUE)
+          if (bs_iter==B) message(paste("\n Done! \n"))
+        
         } else message(paste("\n"))
 
       # in the bootstrap we have to calculate the values for j0 and j1 as the bootstrap
@@ -809,10 +809,9 @@ hellinger.boot.disc <- function(obj, j.max = 10, B = 100, ql = 0.025, qu = 0.975
         # to calculate t0
         # message(paste("Running bootstrap iteration ", bs_iter, " testing for ", j0, " components.\n", sep = ""))
         
-        Sys.sleep(0.5)
         progress(bs_iter, progress.bar = TRUE)
-        #if (bs_iter==B) message(paste("Done! \n"))
-        
+        if (bs_iter==B) message(paste("\n Done! \n"))
+                
       } else message(paste("\n"))
 
       f.n.sqrt.boot <- .get.f.n.sqrt(dat, n.max, N)
