### 2022-06-14 493963c
Update 3_distances_disc.R
