### 2022-06-14 5478f8c
Update 4_distances_cont.R
