### 2022-06-15 3ed6200
Update 3_distances_disc.R
