+++ R/3_distances_disc.R
@@ -371,9 +371,9 @@ L2.boot.disc <- function(obj, j.max = 10, n.inf = 1000, B = 100, ql = 0.025, qu 
         # don't include first iteration as this just uses the original data
         # to calculate t0
         # message(paste("Running bootstrap iteration ", bs_iter, " testing for ", j0, " components.\n", sep = ""))
-          progress(bs_iter, progress.bar = TRUE)
-          Sys.sleep(0.01)
-          if (bs_iter==B) message(paste("\n Done! \n"))
+        progress(bs_iter, progress.bar = TRUE)
+        Sys.sleep(0.01)
+        if (bs_iter==B) message(paste("Done! \n"))
         
         } else message(paste("\n"))
 
@@ -809,10 +809,9 @@ hellinger.boot.disc <- function(obj, j.max = 10, B = 100, ql = 0.025, qu = 0.975
         # don't include first iteration as this just uses the original data
         # to calculate t0
         # message(paste("Running bootstrap iteration ", bs_iter, " testing for ", j0, " components.\n", sep = ""))
-        
         progress(bs_iter, progress.bar = TRUE)
         Sys.sleep(0.01)
-        if (bs_iter==B) message(paste("\n Done! \n"))
+        if (bs_iter==B) message(paste("Done! \n"))
                 
       } else message(paste("\n"))
 
