### 2022-06-15 aa89482
Update 4_distances_cont.R
