### 2022-06-15 d7c31e5
Update 5_lrt_utils.R
