### 2022-06-15 092871b
Update 3_distances_disc.R
