### 2022-06-15 686023b
Update 3_distances_disc.R
