### 2022-06-15 1698386
Update 4_distances_cont.R
