+++ R/4_distances_cont.R
@@ -622,7 +622,7 @@ hellinger.boot.cont <- function(obj, bandwidth, j.max = 10, B = 100, ql = 0.025,
         # don't include first iteration as this just uses the original data
         # to calculate t0
         # message(paste("Running bootstrap iteration ", bs_iter, " testing for ", j0, " components.\n", sep = ""))
-        progress(bs_iter, progress.bar = TRUE)
+        progress(bs_iter, B)
         Sys.sleep(0.01)
         if (bs_iter==B) message(paste("Done! \n"))
         
