### 2022-06-15 ce7eea1
Update 3_distances_disc.R
