### 2022-06-15 b45cc5a
Update 3_distances_disc.R
