### 2022-06-15 cacea43
Update 3_distances_disc.R
