+++ R/3_distances_disc.R
@@ -810,6 +810,7 @@ hellinger.boot.disc <- function(obj, j.max = 10, B = 100, ql = 0.025, qu = 0.975
         # don't include first iteration as this just uses the original data
         # to calculate t0
         # message(paste("Running bootstrap iteration ", bs_iter, " testing for ", j0, " components.\n", sep = ""))
+        if (bs_iter==1) message(paste("\n"))
         progress(bs_iter, B)
         Sys.sleep(0.05)
         if (bs_iter==B) message(paste("Done! \n"))
