### 2022-06-15 11ff3fb
Update 3_distances_disc.R
