### 2022-06-15 93cf9ca
Update 3_distances_disc.R
