+++ R/3_distances_disc.R
@@ -364,6 +364,14 @@ L2.boot.disc <- function(obj, j.max = 10, n.inf = 1000, B = 100, ql = 0.025, qu 
     # counting bootstrap iterations to print progression
     bs_iter <- -1
 
+    ##############
+    ## Basic
+    pb <- progress_bar$new(
+    format = " running [:bar] :percent",
+    total = 100, clear = FALSE, width= 60)
+    ##############
+    
+    
     stat <- function(dat){
       assign("bs_iter", bs_iter + 1, inherits = TRUE)
       if(bs_iter != 0){
@@ -371,10 +379,8 @@ L2.boot.disc <- function(obj, j.max = 10, n.inf = 1000, B = 100, ql = 0.025, qu 
         # don't include first iteration as this just uses the original data
         # to calculate t0
         # message(paste("Running bootstrap iteration ", bs_iter, " testing for ", j0, " components.\n", sep = ""))
-        if (bs_iter==1) message(paste("\n"))
-        Sys.sleep(0.1)
-        progress(bs_iter, B)
-        Sys.sleep(0.1)
+        pb$tick()
+        Sys.sleep(1 / 100)
         if (bs_iter==B) message(paste("Done! \n"))
         
         } else message(paste("\n"))
@@ -803,6 +809,13 @@ hellinger.boot.disc <- function(obj, j.max = 10, B = 100, ql = 0.025, qu = 0.975
     # counting bootstrap iterations to print progression
     bs_iter <- -1
 
+    ##############
+    ## Basic
+    pb <- progress_bar$new(
+    format = " running [:bar] :percent",
+    total = 100, clear = FALSE, width= 60)
+    ##############
+    
     stat <- function(dat){
 
       assign("bs_iter", bs_iter + 1, inherits = TRUE)
@@ -811,10 +824,8 @@ hellinger.boot.disc <- function(obj, j.max = 10, B = 100, ql = 0.025, qu = 0.975
         # don't include first iteration as this just uses the original data
         # to calculate t0
         # message(paste("Running bootstrap iteration ", bs_iter, " testing for ", j0, " components.\n", sep = ""))
-        if (bs_iter==1) message(paste("\n"))
-        Sys.sleep(0.1)
-        progress(bs_iter, B)
-        Sys.sleep(0.1)
+        pb$tick()
+        Sys.sleep(1 / 100)
         if (bs_iter==B) message(paste("Done! \n"))
                 
       } else message(paste("\n"))
