### 2022-06-15 e1ec23d
Update 4_distances_cont.R
