+++ R/4_distances_cont.R
@@ -614,6 +614,13 @@ hellinger.boot.cont <- function(obj, bandwidth, j.max = 10, B = 100, ql = 0.025,
     # counting bootstrap iterations to print progression
     bs_iter <- - 1
 
+    ##############
+    ## Basic
+    pb <- progress_bar$new(
+    format = " running [:bar] :percent",
+    total = 100, clear = FALSE, width= 60)
+    ##############
+    
     stat <- function(dat){
 
       assign("bs_iter", bs_iter + 1, inherits = TRUE)
@@ -622,8 +629,8 @@ hellinger.boot.cont <- function(obj, bandwidth, j.max = 10, B = 100, ql = 0.025,
         # don't include first iteration as this just uses the original data
         # to calculate t0
         # message(paste("Running bootstrap iteration ", bs_iter, " testing for ", j0, " components.\n", sep = ""))
-        progress(bs_iter, B)
-        Sys.sleep(0.01)
+        pb$tick()
+        Sys.sleep(1 / 100)
         if (bs_iter==B) message(paste("Done! \n"))
         
       } else message(paste("\n"))
