+++ R/5_lrt_utils.R
@@ -179,6 +179,14 @@ mix.lrt <- function(obj, j.max = 10, B = 100, quantile = 0.95, control = c(trace
     # counting bootstrap iterations to print progression
     bs_iter <- -1
 
+    ##############
+    ## Basic
+    pb <- progress_bar$new(
+    format = " running [:bar] :percent",
+    total = 100, clear = FALSE, width= 60)
+    ##############
+                           
+                           
     stat <- function(dat){
 
       assign("bs_iter", bs_iter + 1, inherits = TRUE)
@@ -187,9 +195,10 @@ mix.lrt <- function(obj, j.max = 10, B = 100, quantile = 0.95, control = c(trace
         # don't include first iteration as this just uses the original data
         # to calculate t0
         # message(paste("Running bootstrap iteration ", bs_iter, " testing for ", j0, " components.\n", sep = ""))
-        progress(bs_iter, B)
-        Sys.sleep(0.05)
+        pb$tick()
+        Sys.sleep(1 / 100)
         if (bs_iter==B) message(paste("Done! \n"))
+        
         
       } else message(paste("\n"))
 
