### 2022-06-18 #4354 @jbytecode
Dear @yuliadm

- I have just sent a pull request that includes some corrections for the manuscript and the bibtex.
- Please carefully check them. If you are agreed with the corrections, please apply.
- Ping me, when you have done with them.

Thank you in advance.
