### 2022-06-18 #4354 @yuliadm
Dear @jbytecode, thanks for the instructions. I did reread the paper and did not find any mistakes there... Since I will be leaving my registration data on this website (whether sharing it from Git or otherwise), may I at least know why it is required (I do not remember reading about it) and whether there are conflicts of interest from your side. Thanks! Sorry for the question.
