### 2022-06-18 #4354 @jbytecode
@yuliadm - we always ask authors to create an archive of the current form of the software (as stated in https://joss.readthedocs.io/en/latest/editorial_bot.html?highlight=zenodo#setting-the-software-archive and https://joss.readthedocs.io/en/latest/submitting.html?highlight=zenodo#the-review-process).

I only recommended some small changes and there is not a conflict of interest from our side.

Is that explanation clear?
