### 2022-06-24 #4354 @yuliadm
Dear @jbytecode, I tried to follow all of the instructions, so I hope everything is ok:
- here is the tagged release:  https://github.com/yuliadm/mixComp/releases/tag/v0.1.3
- here is the zenodo archive URL: https://doi.org/10.5281/zenodo.6660840
- here is the zenodo archive DOI: 10.5281/zenodo.6660840
Please let me know if I still managed to mess something up, I'll correct it!
Thanks!
