### 2022-06-24 #4354 @yuliadm
@jbytecode I do not think I can do that: they were our master students, but now they are done with their studies and doing something else.
Are there any other options? It is actually written that ORCID is not required...
