### 2022-06-26 #4354 @arfon
@yuliadm – I made some slight changes to your paper to match the current preferred format for denoting equal authorship and corresponding authorship: https://github.com/yuliadm/mixComp/pull/8
